refactor(ai): consolidate document block adapters
