+++ services/ai/providers/anthropic_message_adapter.py
@@ -6,11 +6,13 @@ reject those extras, so providers must pass messages through this adapter before
 sending requests.
 """
 
+import json
 from collections.abc import Iterable
 from typing import cast
 
 from anthropic.types import (
     ContentBlockParam,
+    DocumentBlockParam,
     MessageParam,
     SearchResultBlockParam,
     ToolResultBlockParam,
@@ -29,6 +31,25 @@ type OmniToolResultContentBlockParam = (
     ToolResultContentBlockParam | OmniSearchResultBlockParam
 )
 type AnthropicMessageContent = str | Iterable[OmniContentBlockParam]
+
+
+def extract_text_document(block: DocumentBlockParam) -> str | None:
+    """Convert a text-backed document block into provider-neutral text."""
+    if block.get("type") != "document":
+        return None
+
+    source = block.get("source")
+    if not isinstance(source, dict) or source.get("type") != "text":
+        return None
+
+    data = source.get("data")
+    if not isinstance(data, str) or not data:
+        return None
+
+    title = block.get("title")
+    if isinstance(title, str) and title:
+        return f"Document title: {json.dumps(title, ensure_ascii=False)}\nDocument content:\n{data}"
+    return f"Document content:\n{data}"
 
 
 def build_messages_for_anthropic_api(
