--- services/ai/providers/content_blocks.py
@@ -1,22 +1,0 @@
-import json
-
-from anthropic.types import DocumentBlockParam
-
-
-def extract_text_document(block: DocumentBlockParam) -> str | None:
-    """Convert a text-backed document block into provider-neutral text."""
-    if block.get("type") != "document":
-        return None
-
-    source = block.get("source")
-    if not isinstance(source, dict) or source.get("type") != "text":
-        return None
-
-    data = source.get("data")
-    if not isinstance(data, str) or not data:
-        return None
-
-    title = block.get("title")
-    if isinstance(title, str) and title:
-        return f"Document title: {json.dumps(title, ensure_ascii=False)}\nDocument content:\n{data}"
-    return f"Document content:\n{data}"
